rename pages componets for index.tsx
