+++ project/src/pages/player-pause/index.tsx
