+++ project/src/pages/player/index.tsx
